# Backend/doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Newsletters",
    "version": "1.0.0",
    "description": "Documentation de l'API pour gérer les blogs, utilisateurs, abonnés, commentaires et newsletters."
  },
  "paths": {
    "/newsletters": {
  "get": {
    "summary": "Récupérer toutes les newsletters",
    "tags": ["Newsletters"],
    "responses": {
      "200": {
        "description": "La liste des newsletters",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Newsletter"
              }
            }
          }
        }
      }
    }
  },
  "post": {
    "summary": "Créer une nouvelle newsletter",
    "tags": ["Newsletters"],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "$ref": "#/components/schemas/Newsletter"
          }
        }
      }
    },
    "responses": {
      "201": {
        "description": "Newsletter créée avec succès",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Newsletter"
            }
          }
        }
      },
      "400": {
        "description": "Requête invalide"
      }
    }
  }
},
"/newsletters/{id}": {
  "get": {
    "summary": "Récupérer une newsletter par ID",
    "tags": ["Newsletters"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "200": {
        "description": "Newsletter trouvée",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Newsletter"
            }
          }
        }
      },
      "404": {
        "description": "Newsletter non trouvée"
      }
    }
  },
  "put": {
    "summary": "Mettre à jour une newsletter par ID",
    "tags": ["Newsletters"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "$ref": "#/components/schemas/Newsletter"
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Newsletter mise à jour avec succès",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Newsletter"
            }
          }
        }
      },
      "400": {
        "description": "Requête invalide"
      },
      "404": {
        "description": "Newsletter non trouvée"
      }
    }
  },
  "delete": {
    "summary": "Supprimer une newsletter par ID",
    "tags": ["Newsletters"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "204": {
        "description": "Newsletter supprimée avec succès"
      },
      "404": {
        "description": "Newsletter non trouvée"
      }
    }
  }
}
...
"/users": {
  "get": {
    "summary": "Récupérer tous les utilisateurs",
    "tags": ["Users"],
    "responses": {
      "200": {
        "description": "La liste des utilisateurs",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      }
    }
  },
  "post": {
    "summary": "Créer un nouvel utilisateur",
    "tags": ["Users"],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "$ref": "#/components/schemas/User"
          }
        }
      }
    },
    "responses": {
      "201": {
        "description": "Utilisateur créé avec succès",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "400": {
        "description": "Requête invalide"
      }
    }
  }
},
"/users/{id}": {
  "get": {
    "summary": "Récupérer un utilisateur par ID",
    "tags": ["Users"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "200": {
        "description": "Utilisateur trouvé",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "404": {
        "description": "Utilisateur non trouvé"
      }
    }
  },
  "put": {
    "summary": "Mettre à jour un utilisateur par ID",
    "tags": ["Users"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "$ref": "#/components/schemas/User"
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Utilisateur mis à jour avec succès",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "400": {
        "description": "Requête invalide"
      },
      "404": {
        "description": "Utilisateur non trouvé"
      }
    }
  },
  "delete": {
    "summary": "Supprimer un utilisateur par ID",
    "tags": ["Users"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "204": {
        "description": "Utilisateur supprimé avec succès"
      },
      "404": {
        "description": "Utilisateur non trouvé"
      }
    }
  }
}
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "Nom de l'utilisateur"
          },
          "email": {
            "type": "string",
            "description": "Email de l'utilisateur"
          },
          "password": {
            "type": "string",
            "description": "Mot de passe de l'utilisateur"
          },
          "profilePicture": {
            "type": "string",
            "description": "URL de l'image de profil de l'utilisateur"
          }
        }
      },
...
      "Newsletter": {
        "type": "object",
        "required": ["title", "content", "newsLetterOwnerId"],
        "properties": {
          "title": {
            "type": "string",
            "description": "Titre de la newsletter"
          },
          "content": {
            "type": "string",
            "description": "Contenu de la newsletter"
          },
          "newsLetterOwnerId": {
            "type": "string",
            "description": "ID du propriétaire de la newsletter"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date et heure de création de la newsletter"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date et heure de la dernière mise à jour de la newsletter"
          }
        }
      }
    }
  }
}